refactor: importaciones de pool en controllers
