+++ backend/config/db.js
@@ -1,0 +1,16 @@
+import dotenv from 'dotenv';
+import pkg from 'pg';
+
+dotenv.config();
+
+const { Pool } = pkg;
+
+const pool = new Pool({
+  user: process.env.DB_USER,
+  host: process.env.DB_HOST,
+  database: process.env.DB_NAME,
+  password: process.env.DB_PASSWORD,
+  port: process.env.DB_PORT
+});
+
+export default pool;
